soc-qcom-sc8180x: remove nonexistent package from depends (MR 6019)
The qrtr-openrc package was removed in d3b53c3e624e66e322d2549f69fb11ccc53ba287
but somehow still left in depends of this package.

Fix the errors like:

  ERROR: unable to select packages:
    qrtr-openrc (no such package):
      required by: soc-qcom-sc8180x-nonfree-firmware-openrc-1-r1[qrtr-openrc]

This is the one and only mention of qrtr-openrc package in pmaports.

P.S. It's very telling how many users sc8180x has, if this was unnoticed
for 8 months.

Signed-off-by: Alexey Min <alexey.min@gmail.com>