main/postmarketos-base-ui-gnome: add some gst packages to pmb_recommends
The current OOTB Gstreamer experience on Gnome DEs can be disappointing as
a lot of default codecs and elements are missing. This is often not obvious
for users as Gst in various cases has less optimal fallbacks, easily leaving
the impression that either an app, drivers or hardware are at fault.

Add the most important packages to pmb_recommends so users have a better time.

Notes:
 - Many images already ship ffmpeg-libavcodec, so this shouldn't worsen the
   legal situation.
 - '-bad' is usually already pulled in indirectly while '-good' at least for
   GS variants is already required. Both are strongly recommended for apps
   like Showtime, Snapshot and many more mobile-friendly apps from the
   ecosystem.
 - '-bad' really isn't bad any more these days and various elements in it
   are used by default if available. That is, many elements there should have
   long moved to '-good'.