samsung-loganrelte: boot mainline (MR 6142)
This replaces the downstream kernel with a near-to-mainline kernel, and also adds a msm8930-specific mainline package. While this removes some functionality (screen, touchscreen), this makes the boot more stable (boots no matter if charging or not). To flash this, you first need to flash lk2nd, and then flash to the userdata partition with pmbootstrap flasher flash_rootfs --partition userdata. You might need to wipe the system partition. The boot finishes, but to get USB networking, you might need to dis- and re-connect the device ca. 10 seconds after (re-)booting. SSH and internet-over-USB works fully.