main/postmarketos-base-ui: add universal udev rule for udmabuf (MR 6163)
The udmabuf allocator enjoys increasing popularity, not only being
used by libcameras swISP and some virtualisation apps, but since Mesa
commit d74ea2c1 even in llvmpipe - and more projects plan adoption,
such as Gstreamer.

Thus let's ensure it's universally available for UI users if the
kernel supports it.

As there are DEs that don't support the "uaccess" tag, such as SXMO,
and "legacy" groups are still used in pmOS anyway, use the later.