soc-qcom-sdm845: remove dependency on alsa-utils (MR 6169)
Suddenly apk wants to install other ucm pkgs that conflict with
soc-qcom-sdm845-ucm, like the soc-qcom-msm8996-ucm pkg. Later when the
sdm845-ucm is installed it conflicts with the previously installed (and
incorrect) ucm pkg.

This SOC pkg depends on alsa-utils from 5dc31f3f, but the
call_audio_idle_suspend_workaround.sh that was added in that commit
doesn't actually seem to be using apps from alsa-utils, so maybe it's
safe to remove that dependency and fix an issue where apk will sometimes
install the wrong alsa-ucm-conf pkg for a device...

[ci:skip-build]: already built successfully in CI