linux-next: enable PMEM (MR 6228)
PMEM aka persistent memory enables a region of memory to be treated as a
block device. This is particularly useful as a way for EFI bootloaders
to retrieve a full OS image (via HTTP) and make it accessible to the OS.

Support for this is making its way to U-Boot: https://lore.kernel.org/u-boot/20250203105912.196654-1-sughosh.ganu@linaro.org/

Let's enable this in postmarketOS so we can enable HTTP booting images
directly from postmarketOS.org.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>
[ci:skip-build]: already built successfully in CI