systemd/phosh: use abuild's systemd splitfunc and launch dbus via socket-activation (MR 6243)
The current -systemd subpackage isn't getting installed due to the
invalid install_if statement. As abuild's fork in pmOS for systemd
specifies a default_systemd() and systemd() function, overriding it
here isn't needed.

The -systemd subpackage contains configuration for gnome-session in
systemd mode.
- By placing a file into systemd/user/gnome-session@phosh.target.d,
  we generate a valid gnome-session@phosh.target.
- The first invokation of gnome-session by phosh-session includes
  the arg `--session phosh`. When gnome-session sees a running
  systemd user manager via dbus, it will signal to systemd to start
  `gnome-session@phosh.target`.

As we switched to tinydm/autologin for phosh on systemd-edge, this
will only work when autologin also tries to load pam_systemd.so.
[ci:skip-build]: already built successfully in CI