device-nokia-n900: fix missing makedepend (MR 6268)
pmbootstrap > 3.3.1 is stricter with the dependencies, it will only
install packages that are in makedepends during builds. Before it also
installed packages from depends.

Add postmarketos-mvcfg to makedepends, as it gets used in build().

postmarketos-mvcfg will be removed in the future (see issue 3580), but
let's unbreak the build first.

Related: pmbootstrap issue 2574
[ci:skip-build]: already built successfully in CI