main/postmarketos-mkinitfs-hook-ci: Handle shellcheck checks
 - Point shellcheck to the files this is sourcing
 - Tell it about the shell in use
 - Drop unused variable (grepped around)

The last change is about how the hooks are ran. The warning shellcheck
was giving was that using `if <command...>; then ...; fi` is preferred,
instead of comparing `$?`.

This is exactly what is now being done, so the branches have been
swapped around, and the command (full path to a hook) is being used
within the if condition.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6335