ci: avoid using alpine:edge on lint CI jobs
It is a lot more prone to breaking than stable, which is the
default. The reason to have them there is also not very clear.
The editor-config has been there since its introduction in
7a8de200dea30e3529ab9f8c5d19f77a6b211ac8. It might be that at the time
it was not in community or in a stable release. The shellcheck-grep
has been there since c5e89ac15f2a465c85b37fffb4d465738f5ec536, which
was already 6 years ago. Probably at the time our tooling was a lot
less integrated and stable in alpine as it is today, and everything
was running edge.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6471