main/postmarketos-base-ui: remove "_" from beginning of split functions
This is inconsistent through the file. In addition, the "_" prefix is
often used when a function is not directly a split function from a
subpackage, but a "private" function that others call. So remove it to
make it clear that these are subpackage split functions

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6375