samsung-coreprimevelte: move from testing to community
The device has seen a decent amount of work put into it and is usable
on a basic level. Many features unfortunately remain unavailable, but
the device fulfills the wiki requirements to be moved to community, so
let's do that to increase visibility.

The PXA1908 firmware package and the U-Boot package are moved together
with the device packages.

[ci:skip-vercheck]

Signed-off-by: Duje Mihanović <duje.mihanovic@skole.hr>
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6409
[ci:skip-build]: already built successfully in CI