systemd/networkmanager: Support Librem5 WiFi Hotspot
This applies a patch from the Purism tree [1] to ensure that the Librem5
can create a WiFi Hotspot.

Adding a hacky patch to NetworkManager is less then ideal, but there
currently isn't a better way to get WiFi Hotspot working. pmOS already
has a fork of NetworkManager so adding one more patch isn't too much
worse. If the NetworkManager fork disapears I think this patch can also
just be dropped, but while we are maintaing NetworkManager we may as
well support the Librem5.

Hopefully Purism develop a cleaner fix for this in the future, as
Librem5 development seems to have started up again.

1: https://source.puri.sm/pureos/packages/network-manager/-/blob/pureos/latest/debian/patches/pureos/wifi-Disable-WPA-PSK-SHA256-in-AP-mode.patch

Resolves: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/issues/2508
Signed-off-by: Alistair Francis <alistair@alistair23.me>
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6421
[ci:skip-build]: already built successfully in CI