device-amlogic-tvbox: disable dhcpd
These devices have an ethernet port, therefore from the wiki:

Do not set the static ip to the network interface and do not run the
DHCP daemon on it. This makes sense for devices with an ethernet
port, like the Raspberry Pi (!1580).

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6435