firmware-shift-otter: add support for 8.0 and 8.1
The currently shipped firmware only worked on early EVT and DVT1
devices, because it is unsigned.

Signed firmware, to enable support for 8.0 was uploaded a long
time ago, but the firmware package was not bumped.

With this bump now all variations of otter are supported:
- EVT
- DVT1
- DVT2
- 8.0 (MP)
- 8.1 (MP)

Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6454
[ci:skip-build]: already built successfully in CI