cross/gcc-x86_64: fix depends line for aarch64
The gcc-x86_64 is only built for aarch64. Set the depends line to
so:libc.musl-aarch64.so.1 as we had it before 0992aef9 ("cross/gcc-*:
upgrade to 14.2.0-r6"). I had set it to so:libc.musl-x86_64.so.1 like
the other cross gccs by accident.

[ci:skip-build]: I'll verify it builds locally before merge.
[ci:skip-vercheck]: version not bumped intentionally.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6501