linux-samsung-a6lte-downstream: disable cameras
Having the camera driver in the kernel causes a hangup in wireplumber
when it tries to open a certain /dev/video* device. These camera devices
are not used on postmarketOS anyway, and some of the config options
required for working camera have already been disabled, so just disable
the rest for good.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6644