device-pine64-pinenote: add partition-table
* add `partition-table.txt` which sets the partition table using
`rkdeveloptool`. Due to lack of support from pmbootstrap, currently has
to be done manually.  Comments are added in `deviceinfo` to make that clear.
* removes from `deviceinfo` `deviceinfo_sd_embed_firmware` as unused.
Due to lack of support from pmbootstrap for flashing u-bout using
pmbootstrap, currently has to be done manually. Comments are added
in `deviceinfo` to make that clear.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6711