device-amlogic-tvbox: rename to device-amlogic-aarch64-tvbox
This port is exclusively for AArch64 TV boxes.  Given that there is an
Amlogic TV box port for Meson8 (ARMv7) in the works, it doesn't make
sense for the AArch64 port to have that generic name anymore.  So rename
it to reflect which Amlogic SoCs are suitable for it.

[ci:skip-vercheck]

Signed-off-by: Ferass El Hafidi <funderscore@postmarketos.org>
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6744