google-sargo: install config for plasma-camera
The google-sargo is not quick enough to handle the default plasma-camera
settings. The resolution is a setting too high so this sets it back to
720p which works well enough.
The codec is set to default but I'm adding it to make sure it's
something we know works when the upstream default changes.

The app still says the codec can't keep up but we can't get much better
out of it currently.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6791