device-qcom-msm8953: add wakeup keys for Lenovo ThinkSmart View
This forces the volume keys to wake up the screen.

For now, this device has no other method of waking up the screen if no
external input methods are attached.

On the vendor's Android OS the proximity sensor wakes up
the screen, but it doesn't work on pmOS yet.

The touchscreen is disabled when the screen is blanked, so we cannot use
it to wake up the device.

Signed-off-by: Felix Kaechele <felix@kaechele.ca>
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6832
[ci:skip-build]: already built successfully in CI