u-boot-amlogic: upgrade to v2026.04, switch to gxlimg
The reasoning behind switching to gxlimg for image creation is that the
official Amlogic tools are not only proprietary (which makes them
undesirable in the first place), but they are also x86_64-only, which
increases maintenance burden as hacks have to be implemented for them to
work on non-x86_64 systems. They are also not officially supported, are
buggy, and tend to depend on glibc.

gxlimg has been packaged in Alpine Linux for a very long while, and the
intention behind packaging in Alpine has always been that it will be
used for the u-boot package going forward.

The main disadvantage of gxlimg is that it does not yet support
compressing the u-boot binary in the FIP, but we don't use that feature
in the proprietary tools anyway.

Signed-off-by: Ferass El Hafidi <funderscore@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6900>