ci: move check for package counts to where it is used
common is otherwise over-used with things with very little
consumers. So reduce the list by moving the
get_changed_packages_sanity_check to check_changed_aports_versions,
which is the only user. In the process, also rename it to make it more
clear what it's used for: check_package_count

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6963>