device-fairphone-fp3: modem, audio & qbootctl
While the modem has worked for quite a while on mainline, for some
reason some dependencies were never added to the device package.

And add the firmware for the audio amplifier on Fairphone 3 since it's
also supposed to work now.

And install qbootctl (and enable it by default) to mark the current slot
as successful to not get stuck in the bootloader after a couple of
reboots.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7010>