systemd/postmarketos-base-systemd: add missing preset config
The trigger runs `systemctl preset cups.service`, but then if it runs
preset on the other two *afterwards* it will disable them (because no
explicit preset and we default to disable). This helps ensure they're
kept enabled.

The same seems to also happen for NetworkManager-wait-online and the
other things here, leading to them being accidentally disabled.

This also removes cups.service, it should be socket-activatable now that
cups.socket is enabled.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7029>