soc-qcom-sc7280: add 81voltd daemon as dependency
This daemon configures the modem manager bearers. And
while, due to audio not yet working, this might not yet
be that interesting for VoLTE/VoNR, this configures the
XTRA source and hence increases the chance of a working
GPS with A-GPS and a quick location fix.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7044>