device-samsung-p4note: adjust sensor mount matrix for iio-sensor-proxy
There is currently two different ways to handle sensors, the Linux way
which is taking earth as a starting point and then drawing a vector
through the device, as well as the Windows way which takes the device as
a starting point and drawing a vector towards earth. iio-sensor-proxy
decided to go the Windows way, therefor the configuration needs to be
flipped for the readings to produce the expected result.

I created an issue with iio-sensor-proxy that goes into a bit more
details, see #375.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7085>