CI: Don't check kconfig for archived kernels
It doesn't make any sense to require people to fix the kconfig just
because they want to move a device into archived or fix some minor issue
in an archived device. So, let's not enforce this on archived devices
since that's supposed to be where broken devices go.

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7338>