device-fairphone-fp6: update kernel package name & remove append_dtb
The kernel package has been renamed, so update the dependency
accordingly.

And remove append_dtb, this was actually used with boot-deploy < 0.20.0
because deviceinfo_header_version="2" was ignored and therefore v0 was
used, which requires appended dtb. Now that boot-deploy is actually
passing this parameter, we do not have to append the dtb to the kernel
image anymore, because the dtb is passed in a separate field.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7368>