main/gcc4: upgrade to 4.9.4, fix build
This upgrade GCC 4 to 4.9.4, the last release in the GCC 4 series. This
lets us drop one of the patches, which was included in 4.9.4.

We need to get rid of 3 flags that are set globally for compiling such
an old version of GCC (it doesn't know two of them, and
-Werror=format-security cannot be used with GCC).

Signed-off-by: Jens Reidel <adrian@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7423>