main/postmarketos-base-ui-gnome: enable fractional scaling
Makes it much easier to use plain gnome-shell on tablets or even phones
and is already enabled on GNOME Mobile.

@rmader remebers upstream not enabling this by default because of bugs
with XWayland, but since XWayland applications are much more rare on
mobile phones, plus it's still 100% by default and a fractional scale
has to be enabled manually.

> Makes mutter default to layout logical monitors in a logical pixel
> coordinate space, while scaling monitor framebuffers instead of window
> content, to manage HiDPI monitors. Does not require a restart.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7529>